requirements.txt
# Core
numpy>=1.24,<2
protobuf
grpcio-tools
tqdm
pillow
opencv-python
scipy

# Waymo TFRecord parsing
tensorflow
# waymo-open-dataset-tf-2-12-0 installed separately in Dockerfile (needs --no-build-isolation)

# Detection / Tracking
ultralytics
lap>=0.5.12

# Depth / 3DGS (CPU by default; replace with CUDA wheels on server as needed)
torch
torchvision
torchaudio
transformers

# Point cloud processing
open3d
